improved remove custom scripts for NumPy and SciPy

+++ scripts/bench/python3/scipylib.remove.ps1
@@ -1,1 +1,14 @@
+$pythonDir = App-Dir "Bench.Python3"
+$pip = App-Exe "Bench.Python3.Pip"
+
+if (!(Test-Path $pythonDir))
+{
+    return
+}
+if (!(Test-Path $pip))
+{
+    Write-Error "PIP 3 not found."
+    return
+}
+
 pip uninstall -y scipy
